backend/paji/server/app/flask_extensions/routes_ext/routes/internal/get_all_routes.py
import flask.views

from paji.server.domain import use_cases


class GetAllRoutesView(flask.views.MethodView):

    def __init__(self, get_all_routes_use_case: use_cases.GetAllRoutesUseCase):
        self._get_all_routes_use_case = get_all_routes_use_case

    def get(self):
        return flask.jsonify([
            route.serialize() for route in self._get_all_routes_use_case.execute()
        ])
