backend/requirements-dev.txt
-r requirements.txt
pytest==6.0.2
pytest-mypy==0.7.0
pytest-cov==2.10.1
pytest-flask==1.0.0
